journal_rag/pyproject.toml

[project]
name = "journal-rag"
version = "0.1.0"
description = "A RAG application for processing journal entries with Ollama and Chainlit."
authors = [{"name" = "Your Name", "email" = "you@example.com"}]
dependencies = [
    "llama-index>=0.10.0",
    "chainlit>=1.0.0",
    "python-dotenv>=1.0.0",
    "llama-index-llms-ollama>=0.1.2",
    "llama-index-embeddings-ollama>=0.1.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
